test: split PDF visual equation ownership coverage
